# docs/openapi.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "Agni Dashboard API",
      "version": "1.0.2",
      "description": "Interactive API docs for the Agni Recon Dashboard backend."
    },
    "servers": [
      { "url": "/", "description": "Same-origin" }
    ],
    "paths": {
      "/health": {
        "get": {
          "summary": "Health check",
          "responses": {
            "200": { "description": "OK" }
          }
        }
      },
      "/auth/login": {
        "post": {
          "summary": "Login (get JWT)",
          "requestBody": {
            "required": true,
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } } }
          },
          "responses": {
            "200": { "description": "JWT token", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } } } },
            "401": { "description": "Invalid credentials" }
          },
          "security": []
        }
      },
      "/api/stats": {
        "get": { "summary": "Dashboard stats", "security": [{ "bearerAuth": [] }], "responses": { "200": { "description": "OK" } } }
      },
      "/api/_debug/db": {
        "get": { "summary": "Debug DB name", "responses": { "200": { "description": "OK" } } }
      },
      "/api/recent": {
        "get": {
          "summary": "Recent subdomain changes",
          "security": [{ "bearerAuth": [] }],
          "parameters": [
            { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 200 } }
          ],
          "responses": { "200": { "description": "OK" } }
        }
      },
      "/api/domains": {
        "get": {
          "summary": "List domains",
          "security": [{ "bearerAuth": [] }],
          "parameters": [
            { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1 } },
            { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 200 } }
          ],
          "responses": { "200": { "description": "OK" } }
        }
      },
      "/api/domains/{domain}/subdomains": {
        "get": {
          "summary": "List subdomains for a domain",
          "security": [{ "bearerAuth": [] }],
          "parameters": [
            { "name": "domain", "in": "path", "required": true, "schema": { "type": "string" } },
            { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1 } },
            { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 1000 } },
            { "name": "is_alive", "in": "query", "schema": { "type": "string", "enum": ["true", "false"] } },
            { "name": "has_takeover", "in": "query", "schema": { "type": "string", "enum": ["true", "false"] } }
          ],
          "responses": { "200": { "description": "OK" } }
        }
      },
      "/api/subdomains": {
        "get": {
          "summary": "Search subdomains",
          "security": [{ "bearerAuth": [] }],
          "parameters": [
            { "name": "domain", "in": "query", "schema": { "type": "string" } },
            { "name": "subdomain", "in": "query", "schema": { "type": "string" } },
            { "name": "is_alive", "in": "query", "schema": { "type": "string", "enum": ["true", "false"] } },
            { "name": "has_takeover", "in": "query", "schema": { "type": "string", "enum": ["true", "false"] } },
            { "name": "status_code", "in": "query", "schema": { "type": "integer" } },
            { "name": "q", "in": "query", "schema": { "type": "string" } },
            { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1 } },
            { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 1000 } }
          ],
          "responses": { "200": { "description": "OK" } }
        }
      },
      "/api/takeovers": {
        "get": {
          "summary": "List subdomains with takeover",
          "security": [{ "bearerAuth": [] }],
          "parameters": [
            { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1 } },
            { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 1000 } }
          ],
          "responses": { "200": { "description": "OK" } }
        }
      },
      "/api/export/subdomains": {
        "get": {
          "summary": "Export subdomains as TXT",
          "security": [{ "bearerAuth": [] }],
          "parameters": [
            { "name": "domain", "in": "query", "schema": { "type": "string" } },
            { "name": "is_alive", "in": "query", "schema": { "type": "string", "enum": ["true", "false"] } },
            { "name": "has_takeover", "in": "query", "schema": { "type": "string", "enum": ["true", "false"] } }
          ],
          "responses": { "200": { "description": "OK" } }
        }
      },
      "/api/export/takeovers": {
        "get": {
          "summary": "Export takeovers as TXT",
          "security": [{ "bearerAuth": [] }],
          "responses": { "200": { "description": "OK" } }
        }
      }
    },
    "components": {
      "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" } },
      "schemas": {
        "LoginRequest": { "type": "object", "required": ["username", "password"], "properties": { "username": { "type": "string" }, "password": { "type": "string" } } },
        "AuthResponse": { "type": "object", "properties": { "token": { "type": "string" } } }
      }
    }
  }